helm-base-versions/requirements.txt
azure-cosmos==4.7.0
azure-identity==1.17.1
